tests/test_net_schema.py

from net_schema import (  # replace with the actual name of your module
    definitions,
    init_json_schema_validator,
    init_main_schema,
    init_validator,
)


def test_init_main_schema():
    schema_file = "fixtures/host_vars"  # replace with a path to a test schema file
    main_schema, main_resource = init_main_schema("fixtures/schema.json")

    # Assertions: replace with the actual conditions you're testing for
    assert main_schema is not None
    assert main_resource is not None


def test_load_defs():
    main_id = "http://packetcoders.io/schemas/main"
    main_resource = (
        "your_main_resource"  # replace with an instance of your main resource
    )
    input_def_path = "path/to/your/test/input_def_path"  # replace with a path to a test input definition path
    registry = definitions(main_id, main_resource, input_def_path)

    # Assertions: replace with the actual conditions you're testing for
    assert registry is not None


def test_init_validator():
    main_schema = "your_main_schema"  # replace with an instance of your main schema
    registry = "your_registry"  # replace with an instance of your registry
    validator = init_validator(main_schema, registry)

    # Assertions: replace with the actual conditions you're testing for
    assert validator is not None


def test_init_json_schema_validator():
    schema_path = (
        "path/to/your/test/schema_file"  # replace with a path to a test schema file
    )
    input_def_path = "path/to/your/test/input_def_path"  # replace with a path to a test input definition path
    validator = init_json_schema_validator(schema_path, input_def_path)

    # Assertions: replace with the actual conditions you're testing for
    assert validator is not None
